Move tool to the tools folder
